Restaurant_Staff_Manager/Prototype_01/requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
email-validator
WTForms-SQLAlchemy 
python-dotenv
Flask-Mail
gunicorn  
psycopg2-binary

# AI Forecasting & Data Handling
pandas
prophet



